Proxy element models moved to separate file.
