+++ backend/app/models/__init__.py
@@ -1,3 +1,4 @@
 from .element import *
+from .proxy_elements import *
 from .selector import *
 from .event import *
